Fix no previous tags being seen after checkout
The checkout action, by default, does a shallow clone and only has the ref that triggered the build. Now it will clone the entire git history. This might be slightly slower but at least it'll see the previous tags. When it wasn't seeing the other tags, it always thought that it was the very first version.