Switch from ZonedDateTime to Instant
This is because there is no native codec for ZonedDateTime with MongoDB. Looking at the data mapping and type conversion table from the documentation (https://docs.spring.io/spring-data/mongodb/docs/current/reference/html/#mapping-conversion), it appears that Instant is natively supported. Double-checking what Instant really is (https://stackoverflow.com/a/32443004/1687436), it seems like it's a pretty useful thing. We can just use an Instant, and it will fall upon the client to translate the UTC time to the user's local time zone.