Changed the /city-cis-intercom/verify-user endpoint
The endpoint will now take the username and password as query parameters instead of having an entire schema object for it to be sent as the request body. The endpoint will now also return a UserAccountDto object. This will be cached by the City for any requests that might require information about the user to complete. For instance, creating a new message will require the ID of the user to be known. However, the City is only aware of the username of the user is known during a request. The City will now obtain account info for any user who makes a request. This info will be retained and can be used by any services that need it.