Fix NPE when performing operations on the City members array
When the City document is created, it needs to be persisted with an instance of an ArrayList. Otherwise, the members property isn't even present in the document. Upon retrieval, if you try to operate on the members array, it'll be null. Now Cities are created with an existing List type.