Fixed Tower creation bug
When creating a Tower, the person who created it should be the admin of that tower. I forgot to add this line (I was changing a lot of things at once). This also prevents someone from supplying their own value for the adminId field and creating a Tower that they're the only member of but not the administrator of.