Fix the CIS forgetting the members of a City
This fixes a bug introduced in #27. Something I was aware of with Heroku apps is that [after 30 minutes, they go to sleep](https://blog.heroku.com/app_sleeping_on_heroku) (more details [here](https://devcenter.heroku.com/articles/free-dyno-hours#dyno-sleeping)). Something I wasn't aware of is that they have no persistent file storage. This means that whenever the City service writes its IDENTITY file, that file is lost when the dyno goes to sleep. When the City wakes up again, it sees that it has no IDENTITY file and contacts the [`/city-cis-intercom/register-city`](https://beacon-cis-main-staging.herokuapp.com/swagger-ui/index.html#/City%20Interservice%20Communications/registerCity) CIS endpoint. The CIS checks if there is a City that already exists with the base URL that the "new" City has provided. If there is, it simply sets a new password for that City but lets the ID remain the same. I didn't know this was the case until I had the City members property get initialized every time a new City is registered. What I didn't realize is that this was wiping out any _existing_ array. Now there's a check in case this City already exists and has an array of members.