Remove placeholder endpoints and start building out invite-related endpoints
For a user to redeem an invitation code, they'll need to send it to the CIS since that's what manages all of them. The CIS doesn't have a line of communication to the Cities, so it's not like the CIS can tell the City to add a member to one of its Towers. As a quick fix that gets around this, I think what we'll have to do is have a CIS endpoint that returns info about the relevant City for an invitation code.

Then, the client has to follow up and send the join requests to that specific City. The City will hit a CIS endpoint that makes sure the invite is usable and adds one "use" to it (increments the uses field by 1). Then, the City will go through the same "join and let the CIS know about any new City members" logic that it does when a user joins a tower manually using the API.