Point the production City instance at the QA CIS instance
This was done temporarily just so that we had two City instances running on separate machines with separate databases for demonstration purposes. By default, the production City is configured to point at a production CIS instance. However, that'd mess up identity stuff so we're gonna point it at our existing CIS that we'll be using for the demo.

This should be reverted after we're done with the demo. Created #39 to keep track of this.