Make the /attachments resource handler conditional
Now it will only be created if we're using local storage to host user-uploaded content. If we're using a cloud solution, there's no need to add the resource handler as it'll have no resources to provide. This also makes it so that we can use @WebMvcTest and not have it freak out that it can't find an instance of LocalStorageProperties.