Updated the endpoint for updating a pfp
It now deletes the previous picture from DigitalOcean Spaces. It also now uses a unique ID in the file name each time so that the URL is different each time. This fixes two issues. First, the CDN cache won't be referring to the old file. Second, the frontend will see that there's a new URL so the components that depend on that string will update.